PubSub from same connection was contending on lock checking closed state.
