+++ .travis.yml
@@ -1,6 +1,7 @@
 language: go
 sudo: false
 go:
+- 1.13.x
 - 1.12.x
 - 1.11.x
 env:
