+++ go.mod
@@ -1,7 +1,9 @@
 module github.com/nats-io/nats.go
 
 require (
+	github.com/golang/protobuf v1.3.2
 	github.com/nats-io/jwt v0.3.0
+	github.com/nats-io/nats-server/v2 v2.1.0
 	github.com/nats-io/nkeys v0.1.0
 	github.com/nats-io/nuid v1.0.1
 )
