+++ test/bench_test.go
@@ -70,10 +70,6 @@ func BenchmarkPubSubSpeed(b *testing.B) {
 		if err := nc.Publish("foo", msg); err != nil {
 			b.Fatalf("Error in benchmark during Publish: %v\n", err)
 		}
-		// Don't overrun ourselves and be a slow consumer, server will cut us off
-		if int32(i)-atomic.LoadInt32(&received) > 32768 {
-			time.Sleep(100 * time.Nanosecond)
-		}
 	}
 
 	// Make sure they are all processed.
